+++ packages/create-cloudflare/templates/astro/pages/templates/js/wrangler.jsonc
