+++ packages/create-cloudflare/templates/astro/pages/templates/ts/src/env.d.ts
