+++ packages/create-cloudflare/templates/astro/pages/templates/ts/worker-configuration.d.ts
