+++ packages/create-cloudflare/templates/astro/pages/templates/ts/wrangler.jsonc
