+++ packages/create-cloudflare/templates/astro/workers/templates/js/public/.assetsignore
