+++ packages/create-cloudflare/templates/astro/workers/templates/js/wrangler.jsonc
