+++ packages/create-cloudflare/templates/astro/workers/templates/ts/public/.assetsignore
