+++ packages/create-cloudflare/templates/astro/workers/templates/ts/src/env.d.ts
