+++ packages/create-cloudflare/templates/astro/workers/templates/ts/worker-configuration.d.ts
