+++ packages/create-cloudflare/templates/astro/workers/templates/ts/wrangler.jsonc
